requirements.txt
python-telegram-bot==20.7
openai==1.3.7
python-dotenv==1.0.0
sqlalchemy==1.4.53
alembic==1.12.1
pytz==2023.3
pandas==2.1.3
openpyxl==3.1.2
Pillow==10.1.0
matplotlib>=3.5.0
pytest>=7.4.0
pytest-asyncio>=0.23.0
psutil>=5.9.0
